bash: show name of remote branch in prompt
When a remote branch is checked out (not creating a local branch
tracking it), `rev-parse` just prints `HEAD`. This tries to resolve the
remote branch, and prints that if available.

This process can take an unacceptable amount of time for loading a shell
prompt when run on large repos. We address this problem by timing out
early, setting the timeout to just how long is needed for certain
benchmark repos.